+++ test/Scripts.Integration.Test/build-project.ps1
@@ -25,5 +25,12 @@ else
     RunUnityCustom $unityPath $unityArgs
 }
 
+if ($Platform -eq "Android-Export")
+{
+    # See test/Scripts.Integration.Test/gradle/README.md
+    $gradleVersion = $UnityVersion.StartsWith("2019") ? "v5.1.1" : "v6.1.1"
+    Copy-Item -Force -Recurse "$IntegrationScriptsPath/gradle/$gradleVersion/*" -Destination $outputPath
+}
+
 Write-Host "Project built successfully" -ForegroundColor Green
 Get-ChildItem $NewProjectBuildPath